Implementing Spring Security #16 [added big int(Long) to User, Role and Privilege]

+++ src/main/java/com/mia/bookstore/model/Privilege.java
@@ -15,7 +15,7 @@ public class Privilege {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer privilegeId;
+    private Long privilegeId;
 
     private String name;
