+++ src/main/java/com/mia/bookstore/repository/PrivilegeRepository.java
@@ -3,5 +3,5 @@ package com.mia.bookstore.repository;
 import com.mia.bookstore.model.Privilege;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrivilegeRepository extends JpaRepository<Privilege, Integer> {
+public interface PrivilegeRepository extends JpaRepository<Privilege, Long> {
 }
